removed binance alpha main page promo; replaced with s5 start
