+++ src/lib/domains/splashpage/components/index.svelte
@@ -4,11 +4,10 @@
   import S4StartBanner from '$lib/domains/splashpage/components/Banner/S4StartBanner.svelte';
   import type { Slide } from '$lib/domains/splashpage/components/Banner/types';
   import FeaturedCampaigns from '$shared/components/FeaturedCampaigns/FeaturedCampaigns.svelte';
-  import { BinanceAlphaModal } from '$shared/components/Modals/BinanceAlphaModal';
   import { classNames } from '$shared/utils/classNames';
 
-  import BinanceAlphaBanner from './Banner/BinanceAlphaBanner.svelte';
   import S3ClaimBanner from './Banner/S3ClaimBanner.svelte';
+  import S5StartBanner from './Banner/S5StartBanner.svelte';
   import Factions from './Factions/Factions.svelte';
   import { TrailblazerGuide } from './Guide';
   import HeroSection from './HeroSection/HeroSection.svelte';
@@ -45,7 +44,7 @@
   <HeroSection />
   <div class={separator40pxClasses} />
 
-  <BinanceAlphaBanner />
+  <S5StartBanner />
 
   <div class={separator40pxClasses} />
 
@@ -63,5 +62,3 @@
 
   <EcosystemSection />
 </div>
-
-<BinanceAlphaModal />
