--- src/lib/shared/components/Modals/BinanceAlphaModal/index.ts
@@ -1,1 +1,0 @@
-export { default as BinanceAlphaModal } from './BinanceAlpha.modal.svelte';
